Fixed token auth issues and locked gems

+++ db/migrate/20160811215429_create_ingredients.rb
@@ -2,7 +2,7 @@ class CreateIngredients < ActiveRecord::Migration[5.0]
   def change
     create_table :ingredients do |t|
       t.string :name
-      t.integer :percentage
+      t.decimal :percentage
       t.references :recipe, foreign_key: true
 
       t.timestamps
